week_3/day_3/chatbot_project/requirements.txt
flask==2.3.3
openai==0.28.1
python-dotenv==1.0.0